Move OTC tests into main auth test suite
